refactor login component and auth service; remove login service, enhance error handling, and streamline userId management
