+++ src/app/services/auth.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Router } from '@angular/router';
+import { UserService } from './user.service';
 
 interface LoginResponse {
   token: string;
@@ -11,29 +10,45 @@ interface LoginResponse {
   providedIn: 'root'
 })
 export class AuthService {
-  constructor(private http: HttpClient, private router: Router) {}
+  userService = inject(UserService);  
+  urlApi = 'http://localhost:8000/api';
 
-  async login(email: string, password: string): Promise<LoginResponse> {
-    const response = await fetch('/api/login', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({ email, password })
-    });
-
-    if (!response.ok) {
-      throw new Error('Login failed');
+  async login(email: string, password: string): Promise<any> {
+    if (!email || !password) {
+      throw new Error('Both email and password are required.');
     }
 
-    const data = await response.json();
-    localStorage.setItem('userId', data.user.id);
-    return data;
+    if (password.length < 6) {
+      throw new Error('Password must be at least 6 characters long.');
+    }
+
+    const loginData = { email, password };
+
+    try {
+      const response = await fetch(`${this.urlApi}/login`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(loginData),
+      });
+
+      if (!response.ok) {
+        const errorData = await response.json();
+        throw new Error(errorData.errorMessage || 'Login failed.');
+      }
+
+      const data = await response.json();
+      localStorage.setItem('userId', data.user.id);
+      this.userService.loadUser(data.user.id);
+      return data;
+    } catch (error: any) {
+      throw new Error(error.message || 'An unexpected error occurred. Please try again later.');
+    }
   }
 
   logout() {
     localStorage.removeItem('userId');
-    this.router.navigate(['/login']);
   }
 
   isLoggedIn(): boolean {
